Add WidgetPod::set_origin, soft-deprecate set_layout_rect
As outlined in https://github.com/linebender/druid/issues/1242.

Allowing a parent to give a widget a size different from the size
it had computed in its `layout` method had the potential to cause
subtle breakage. With this change, we now always use the size
returned by the widget's layout method, and expect the container
to set the origin only, using WidgetPod::set_origin.

This does not remove or (yet) officially deprecate
WidgetPod::set_layout_rect; I am curious to see if there is any
existing code where a widget was being given a size different
from what it had returned, and deprecating this method would
cause us to fail in CI. I expect to deprecate it before 0.7.0, though.

This is motivated by the work on baselines; baseline offsets are one
of the things that would be subtly broken if the child's chosen
size and the final size differed.